Fix #1: Code refactoring
