+++ binary_search_tree.rb
@@ -1,12 +1,4 @@
-class Node
-  attr_accessor :value, :left_node, :right_node
-
-  def initialize(value)
-    @value = value
-    @left_node = nil
-    @right_node = nil
-  end
-end
+require_relative "node"
 
 class BinarySearchTree
   attr_accessor :root
@@ -16,8 +8,9 @@ class BinarySearchTree
   end
 
   def add_node(element)
+    new_node = Node.new(element)
     if @root.nil?
-      @root = Node.new(element)
+      @root = new_node
     else
       parent = nil
       current = @root
@@ -30,9 +23,9 @@ class BinarySearchTree
         end
       end
       if element < parent.value
-        parent.left_node = Node.new(element)
+        parent.left_node = new_node
       else
-        parent.right_node = Node.new(element)
+        parent.right_node = new_node
       end
     end
   end
@@ -44,8 +37,8 @@ class BinarySearchTree
   end
 
   def largest_element(node = @root)
-    unless node.nil?
-      unless node.right_node.nil?
+    if node
+      if node.right_node
         largest_element(node.right_node)
       else
         return node.value
@@ -56,8 +49,8 @@ class BinarySearchTree
   end
 
   def smallest_element(node = @root)
-    unless node.nil?
-      unless node.left_node.nil?
+    if node
+      if node.left_node
         smallest_element(node.left_node)
       else
         return node.value
@@ -68,7 +61,7 @@ class BinarySearchTree
   end
 
   def in_order_traversal(node = @root)
-    unless node.nil?
+    if node
       in_order_traversal(node.left_node)
       puts " #{node.value} "
       in_order_traversal(node.right_node)
@@ -76,7 +69,7 @@ class BinarySearchTree
   end
 
   def pre_order_traversal(node = @root)
-    unless node.nil?
+    if node
       puts " #{node.value} "
       pre_order_traversal(node.left_node)
       pre_order_traversal(node.right_node)
@@ -84,7 +77,7 @@ class BinarySearchTree
   end
 
   def post_order_traversal(node = @root)
-    unless node.nil?
+    if node
       post_order_traversal(node.left_node)
       post_order_traversal(node.right_node)
       puts " #{node.value} "
@@ -93,16 +86,16 @@ class BinarySearchTree
 
   def level_order_traversal
     node = @root
-    unless node.nil?
+    if node
       queue = []
       queue << node
       until queue.empty?
         node = queue.shift
         puts " #{node.value} "
-        unless node.left_node.nil?
+        if node.left_node
           queue << node.left_node
         end
-        unless node.right_node.nil?
+        if node.right_node
           queue << node.right_node
         end
       end
@@ -110,7 +103,7 @@ class BinarySearchTree
   end
 
   def search_element(element, node = @root)
-    unless node.nil?
+    if node
       if element < node.value
         return search_element(element, node.left_node)
       elsif element > node.value
@@ -124,7 +117,7 @@ class BinarySearchTree
   end
 
   def remove_element(element, node = @root)
-    unless node.nil?
+    if node
       if element < node.value
         node.left_node = remove_element(element, node.left_node)
       elsif element > node.value
@@ -152,62 +145,3 @@ class BinarySearchTree
   end
 end
 
-
-tree = BinarySearchTree.new
-
-loop do
-  puts "Binary Search Tree User Interface"
-  puts "1. Add elements into the tree(multiple elements seperated by ,)"
-  puts "2. Print largest element"
-  puts "3. Print smallest element"
-  puts "4. Print Inorder, postorder, level order and preorder traversal"
-  puts "5. Search an Element"
-  puts "6. Remove an Element"
-  puts "Enter quit to stop the program"
-  puts "Enter your choice 1 - 6 or quit"
-  choice = gets.chomp
-  break if choice.upcase == "QUIT"
-
-  case choice
-  when "1"
-    puts "Enter the elements you want in the tree"
-    array = gets.chomp.scan(/\w+/)
-    tree.add_elements(array.map(&:to_i))
-  
-  when "2"
-    puts "Largest Element is #{tree.largest_element}"
-
-  when "3"
-    puts "Smallest Element is #{tree.smallest_element}"
-
-  when "4"
-    puts "In-order Traversal"
-    tree.in_order_traversal
-
-    puts "Pre-order Traversal"
-    tree.pre_order_traversal
-
-    puts "Post-order Traversal"
-    tree.post_order_traversal
-
-    puts "Level-order Traversal"
-    tree.level_order_traversal
-  
-  when "5"
-    puts "Enter an element to search"
-    element = gets.chomp.to_i
-    if tree.search_element(element)
-      puts "#{element} is present in the tree"
-    else
-      puts "#{element} is not present in the tree"
-    end
-
-  when "6"
-    puts "Enter an element to remove"
-    element = gets.chomp.to_i
-    tree.remove(element)
-
-  else
-    puts "Enter valid choice"
-  end
-end
