+++ node.rb
@@ -1,0 +1,9 @@
+class Node
+  attr_accessor :value, :left_node, :right_node
+
+  def initialize(value)
+    @value = value
+    @left_node = nil
+    @right_node = nil
+  end
+end
